core_requirements.txt
irc
sphinx
nosetests

